Tests/LibWeb/Text/expected/css/css-inline-import-initiator-type.txt

Inline @import initiator type: css
